Add Commands for buttons in analyzer control window. Add log autoscroll.

+++ AnalyzerControlApp/PresentationWinForms/Windows/ConnectionSettingsWindow.cs
@@ -112,7 +112,7 @@ namespace PresentationWinForms.Forms
             int baudrate = int.Parse(selectBaudrate.SelectedItem.ToString());
 
             analyzer.Options.PortName = portName;
-            analyzer.Options.Baudrate = (uint)baudrate;
+            analyzer.Options.Baudrate = baudrate;
 
             analyzer.SaveConfiguration("AnalyzerServiceConfiguration");
